+++ src/main/java/com/github/attacktive/msaproduct/product/adapter/outbound/persristence/ProductRepository.java
@@ -1,4 +1,4 @@
-package com.github.attacktive.msaproduct.product.port.outbound.persristence;
+package com.github.attacktive.msaproduct.product.adapter.outbound.persristence;
 
 import java.util.Set;
 
@@ -8,6 +8,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ProductJpaRepository extends JpaRepository<ProductEntity, Long> {
+public interface ProductRepository extends JpaRepository<ProductEntity, Long> {
 	Page<ProductEntity> findAllByIdIn(Set<Long> ids, PageRequest pageRequest);
 }
